# go-backend/docs/openapi/wave1.yaml
openapi: 3.0.3
info:
  title: POS Bengkel Go API (Wave 1)
  version: 0.1.0
servers:
  - url: http://localhost:8081
paths:
...
  /api/v1/cash-management/change/suggest:
    post:
      summary: Suggest optimal cash change breakdown based on available drawer stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                total_due:
                  type: integer
                received:
                  type: array
                  items:
                    type: object
                    properties:
                      denomination_id:
                        type: integer
                      quantity:
                        type: integer
      responses:
        '200':
          description: Change suggestion payload
        '422':
          description: Invalid payload or received amount is insufficient
  /api/v1/cash-management/sale/settle:
    post:
      summary: Settle cash sale payment and persist change transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                total_due:
                  type: integer
                description:
                  type: string
                  nullable: true
                actor_user_id:
                  type: integer
                  nullable: true
                received:
                  type: array
                  items:
                    type: object
                    properties:
                      denomination_id:
                        type: integer
                      quantity:
                        type: integer
      responses:
        '200':
          description: Cash sale settled payload
        '422':
          description: Invalid payload or insufficient change stock
...